vertex ai backend/requirements.txt
google-generativeai
google-adk
fastapi
uvicorn
click
python-dotenv
redis
firebase-admin